pyproject.toml
[tool.poetry]
name = "telegram-ai-handler"
version = "1.0.0"
description = "Production-ready Telegram personal account handler with AI integration"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
python = "^3.11"

[tool.poetry.dependencies]
python = "^3.11"
telethon = "^1.29.3"
pyrogram = "^2.0.106"
python-dotenv = "^1.0.0"
chromadb = "^0.4.15"
sentence-transformers = "^2.2.2"
openai = "^1.3.0"
pypdf = "^3.17.0"
python-docx = "^1.1.0"
textract = "^1.6.5"
tiktoken = "^0.5.1"
aiofiles = "^23.2.1"
asyncio-throttle = "^1.0.2"
cryptography = "^41.0.7"
pillow = "^10.1.0"
numpy = "^1.24.3"
pandas = "^2.0.3"
tqdm = "^4.66.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.7.1"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3